Chisel's Transparent Scaffolding fix
Letting players use the in-dev texture while waiting for Chisel to release the update.

This texture was most recently updated by Drillkus (previously made or edited by tterrag1098).  Used without express permission, but is licensed under either Chisel's GPLv2 or CC BY-NC-SA 4.0, both of which are open source and allow non-profit redistribution (CC requires attribution, which is the first line of this paragraph).